# dapp-verifier/verifier-api/src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Verifier API Swagger Documentation - WIP",
      "version": "1.0.0"
    },
    "schemes": ["http"],
    "servers": [{ "url": "http://localhost:5000/" }],
    "paths": {
      "/group": {
        "post": {
          "tags": ["Group"],
          "summary": "Add a new public identity commitment in group",
          "description": "Send group id and identity commitment to push it in the allowed identity commitments in the group",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRequestSchema"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GroupResponseSchema"
                  }
                }
              }
            },
            "400": { "description": "Bad Request" },
            "500": { "description": "Internal server error" }
          }
        }
      },
      "/group/{id}": {
        "get": {
            "tags": ["Group"],
            "summary": "Get all commitments by group ID",
            "description": "Send group id and get all identity commitments in the group",
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/GroupResponseSchema"
                    }
                  }
                }
              },
              "404": { "description": "Not Found" },
              "500": { "description": "Internal server error" }
            },
            "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "type": "string",
                  "format": "string"
                }
              ]
          }
      },
      "/identity": {
        "post": {
          "tags": ["Identity"],
          "summary": "Create an entry for a verified identity",
          "description": "Store the identity commitment, blockchain address and the zk proof",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdentitySchema"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserIdentityResponseSchema"
                  }
                }
              }
            },
            "400": { "description": "Identity already exists" },
            "500": { "description": "Internal server error" }
          }
        }
      },
      "/identity/commitment/{commitment}": {
        "get": {
            "tags": ["Identity"],
            "summary": "Get stored identity by commitment",
            "description": "Send identity commitment and get blockchain address and zk proof associated with it",
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserIdentityResponseSchema"
                    }
                  }
                }
              },
              "404": { "description": "Not Found" },
              "500": { "description": "Internal server error" }
            },
            "parameters": [
                {
                  "name": "commitment",
                  "in": "path",
                  "required": true,
                  "type": "string",
                  "format": "string"
                }
              ]
          }
      },
      "/identity/blockchainaddress/{bcAddress}": {
        "get": {
            "tags": ["Identity"],
            "summary": "Get all commitments by group ID",
            "description": "Send blockchain address and get identity commitment and zk proof associated with it",
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserIdentityResponseSchema"
                    }
                  }
                }
              },
              "404": { "description": "Not Found" },
              "500": { "description": "Internal server error" }
            },
            "parameters": [
                {
                  "name": "bcAddress",
                  "in": "path",
                  "required": true,
                  "type": "string",
                  "format": "string"
                }
              ]
          }
      }
    },
    "components": {
      "schemas": {
        "GroupRequestSchema": {
            "properties": {
              "groupId": {
                "type": "string",
                "example": "1",
                "required": true
              },
              "identityCommitment": {
                "type": "string",
                "example": "1234567",
                "required": true
              }
            }
        },
        "GroupResponseSchema": {
            "properties": {
              "groupId": {
                "type": "string",
                "example": "1",
                "required": true
              },
              "identityCommitment": {
                "type": "array",
                "example": "['12345','67890']",
                "required": true
              }
            }
        },
        "UserIdentitySchema": {
            "properties": {
                "identityCommitment": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                },
                "blockchainAddress": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                },
                "zkProof": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                }
            }
        },
        "UserIdentityResponseSchema": {
            "properties": {
                "identityCommitment": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                },
                "blockchainAddress": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                },
                "zkProof": {
                    "type": "string",
                    "example": "1234567...",
                    "required": true
                }
            }
        }
      }
    }
  }